KML002 - B, C: Update logic codes

+++ nts.uk/uk.at/at.ctx/schedule/nts.uk.ctx.at.schedule.dom/src/main/java/nts/uk/ctx/at/schedule/dom/budget/schedulevertical/verticalsetting/FormBuilt.java
@@ -1,5 +1,7 @@
 package nts.uk.ctx.at.schedule.dom.budget.schedulevertical.verticalsetting;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import nts.arc.enums.EnumAdaptor;
@@ -22,7 +24,7 @@ public class FormBuilt {
     private int verticalCalItem1;
     
     /* 縦計入力項目 */
-    private int verticalInputItem1;
+    private BigDecimal verticalInputItem1;
     
     /* 設定方法 */
     private SettingMethod settingMethod2;
@@ -30,17 +32,23 @@ public class FormBuilt {
     private int verticalCalItem2;
     
     /* 縦計入力項目 */
-    private int verticalInputItem2;
+    private BigDecimal verticalInputItem2;
     
     /* 演算子区分 */
     private OperatorAtr operatorAtr;
 
-	public static FormBuilt createFromJavaTypeFormBuilt(String companyId, String verticalCalCd, String verticalCalItemId, int settingMethod1, int verticalCalItem1, int verticalInputItem1, 
-    		int settingMethod2, int verticalCalItem2, int verticalInputItem2, int operatorAtr) {
+	public static FormBuilt createFromJavaTypeFormBuilt(String companyId, String verticalCalCd, String verticalCalItemId, int settingMethod1, int verticalCalItem1, 
+			BigDecimal verticalInputItem1, int settingMethod2, int verticalCalItem2, BigDecimal verticalInputItem2, int operatorAtr) {
 
-		return new FormBuilt(companyId, verticalCalCd, 
-				verticalCalItemId, EnumAdaptor.valueOf(settingMethod1, SettingMethod.class),
-				verticalCalItem1, verticalInputItem1, EnumAdaptor.valueOf(settingMethod2, SettingMethod.class),
-				verticalCalItem2, verticalInputItem2, EnumAdaptor.valueOf(operatorAtr, OperatorAtr.class));
+		return new FormBuilt(companyId, 
+				verticalCalCd, 
+				verticalCalItemId, 
+				EnumAdaptor.valueOf(settingMethod1, SettingMethod.class),
+				verticalCalItem1, 
+				verticalInputItem1, 
+				EnumAdaptor.valueOf(settingMethod2, SettingMethod.class),
+				verticalCalItem2, 
+				verticalInputItem2, 
+				EnumAdaptor.valueOf(operatorAtr, OperatorAtr.class));
 	}
 }
